The cli can now compile and execute .rock programs.

+++ bedrock/src/Data/Bedrock/Compile.hs
@@ -1,14 +1,15 @@
 module Data.Bedrock.Compile where
 
+import           Control.Monad                     (when)
 import           System.FilePath
 import           Text.ParserCombinators.Parsec     (parseFromFile)
-import Text.Printf
+import           Text.Printf
 
 --import           Data.Bedrock
 import           Data.Bedrock.EvalApply
 import           Data.Bedrock.Exceptions
 import           Data.Bedrock.GlobalVariables
-import           Data.Bedrock.HPT                 
+import           Data.Bedrock.HPT
 import           Data.Bedrock.Invoke
 import           Data.Bedrock.LLVM                 as LLVM
 import           Data.Bedrock.Parse
@@ -16,11 +17,15 @@ import           Data.Bedrock.PrettyPrint
 import           Data.Bedrock.RegisterIntroduction
 import           Data.Bedrock.Rename
 --import           Data.Bedrock.Simplify
-import           Data.Bedrock.Storage
-import           Data.Bedrock.Storage.Pluggable
-import           Data.Bedrock.Storage.Fixed
-import           Data.Bedrock.NodeSizing
 import           Data.Bedrock
+import           Data.Bedrock.NodeSizing
+import           Data.Bedrock.Storage
+import           Data.Bedrock.Storage.Fixed
+import           Data.Bedrock.Storage.Pluggable
+
+-- Compile options
+type KeepIntermediateFiles = Bool
+type Verbose = Bool
 
 type Pipeline = [Step]
 data Step
@@ -31,8 +36,10 @@ data Step
 infixr 9 :>
 infixr 9 :?>
 
-runPipeline :: String -> Module -> Pipeline -> IO Module
-runPipeline title m0 = worker hpt0 0 m0
+runPipeline :: KeepIntermediateFiles -> Verbose
+            -> String -> Module -> Pipeline -> IO Module
+runPipeline keepIntermediateFiles verbose title m0 =
+    worker hpt0 0 m0
   where
     worker _ _ m [] = return m
     worker hpt n m (step:steps) =
@@ -46,20 +53,26 @@ runPipeline title m0 = worker hpt0 0 m0
             PerformHPT ->
                 worker (runHPT m) n m steps
     runAction n m tag action = do
-        printf "[%d] Running step %s\n" (n::Int) (show tag)
+        when verbose $
+            printf "[%d] Running step %s\n" (n::Int) (show tag)
         let m' = action m
-        writeFile (dstFile n tag) (show $ ppModule m')
+        when keepIntermediateFiles $
+            writeFile (dstFile n tag) (show $ ppModule m')
         return m'
     hpt0 = runHPT m0
     dstFile n tag = title <.> show n <.> tag <.> "rock"
 
 compileFromFile :: FilePath -> IO ()
-compileFromFile path = do
+compileFromFile = compileFromFileWithOpts True True
+
+compileFromFileWithOpts :: KeepIntermediateFiles -> Verbose
+                        -> FilePath -> IO ()
+compileFromFileWithOpts keepIntermediateFiles verbose path = do
     ret <- parseFromFile parseModule path
     case ret of
         Left err -> print err
         Right m  -> do
-            result <- runPipeline base m
+            result <- runPipeline keepIntermediateFiles verbose base m
                 [ "rename"          :> unique
                 , PerformHPT
                 , "no-laziness"     :?> runGen . lowerEvalApply
